[ubuntu] Move build scripts out of github workflow
These are in principle fairly independent of the github action and could be used locally
or as references for local build if needed.

More importantly though, by moving the build scripts out of the workflow file,
we could potentially detect changes in this file in the workflow to trigger updates.
We could also make automatic commits to update this without touching the workflow files.